+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 #### Next
 
+* Fix template caching for multiple formats. [#43](https://github.com/ruby-grape/grape-rabl/pull/43) [@kushkella](https://github.com/kushkella)
 * Your contribution here.
 
 #### v0.4.1